+++ lib/matplotlib/backends/backend_gtk3agg.py
@@ -5,12 +5,12 @@ import six
 import warnings
 
 from matplotlib._pylab_helpers import Gcf
-from ._backend_gtk3agg import (FigureCanvasGTK3Agg,
+from .base_backend_gtk3agg import (FigureCanvasGTK3Agg,
                             new_figure_manager,
                             new_figure_manager_given_figure
                             )
 
-from ._backend_gtk3 import FigureManagerGTK3
+from .base_backend_gtk3 import FigureManagerGTK3
 from .backend_gtk3 import (show,
                            draw_if_interactive,
                            _gtk_cleanup,
