+++ lib/matplotlib/backends/base_backend_gtk.py
