+++ lib/matplotlib/backends/base_backend_gtk3.py
