+++ lib/matplotlib/backends/base_backend_qt4.py
